send ABC to opencl module
